requirements.txt
# LeadSniper Backend Dependencies
# Python 3.10+ required

# Database
supabase>=2.0.0

# Web Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0  # Fast HTML parser for BeautifulSoup
google-search-results>=2.4.0  # SerpAPI for domain hunting

# Email (httpx for async, thread-safe per-request API key)
resend>=0.7.0
httpx>=0.27.0

# Data Validation
pydantic>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# CLI & Logging
rich>=13.0.0

# Async utilities (included with Python, but explicit for clarity)
# asyncio is part of stdlib

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# mypy>=1.0.0
